+++ tests/unit/exportplan/test_helpers.py
@@ -386,6 +386,26 @@ def test_get_current_url_country_required_not_in_check(mock_get_exportplan):
     assert current_url.get('country_required') is None
 
 
+@pytest.mark.parametrize('export_plan_data, expected', [
+    [{'export_commodity_codes': [{'commodity_code': '220850', 'commodity_name': 'Gin'}]}, None],
+    [{'export_commodity_codes': []}, True],
+    [{'export_commodity_codes': None}, True],
+])
+@mock.patch.object(helpers, 'get_exportplan')
+def test_get_current_url_product_required(mock_get_exportplan, export_plan_data, expected):
+    mock_get_exportplan.return_value = export_plan_data
+    current_url = helpers.get_current_url(slug='target-markets-research', export_plan=export_plan_data)
+    assert current_url.get('product_required') == expected
+
+
+@mock.patch.object(helpers, 'get_exportplan')
+def test_get_current_url_product_required_not_in_check(mock_get_exportplan):
+    export_plan_data = {'export_commodity_codes': []}
+    mock_get_exportplan.return_value = export_plan_data
+    current_url = helpers.get_current_url(slug='about-your-business', export_plan=export_plan_data)
+    assert current_url.get('product_required') is None
+
+
 @mock.patch.object(api_client.dataservices, 'get_population_data_by_country')
 def test_get_population_data_by_country(mock_population_data_by_country):
     data = {'country': 'United Kingdom', 'population_data': {'target_population': 10000}}